Refactor routes detection
