+++ packages/now-build-utils/src/detect-builders.ts
@@ -1,5 +1,7 @@
 import minimatch from 'minimatch';
 import { valid as validSemver } from 'semver';
+import { parse as parsePath, extname } from 'path';
+import { Route, Source } from '@now/routing-utils';
 import { PackageJson, Builder, Config, BuilderFunctions } from './types';
 
 interface ErrorResponse {
@@ -17,12 +19,58 @@ interface Options {
     buildCommand?: string | null;
     outputDirectory?: string | null;
   };
+  cleanUrls?: boolean;
+  trailingSlash?: boolean;
+  featHandleMiss?: boolean;
 }
 
 // We need to sort the file paths by alphabet to make
 // sure the routes stay in the same order e.g. for deduping
 export function sortFiles(fileA: string, fileB: string) {
   return fileA.localeCompare(fileB);
+}
+
+export function detectApiExtensions(builders: Builder[]): Set<string> {
+  return new Set<string>(
+    builders
+      .filter(
+        b =>
+          b.config && b.config.zeroConfig && b.src && b.src.startsWith('api/')
+      )
+      .map(b => extname(b.src))
+      .filter(Boolean)
+  );
+}
+
+export function detectApiDirectory(builders: Builder[]): string | null {
+  // TODO: We eventually want to save the api directory to
+  // builder.config.apiDirectory so it is only detected once
+  const found = builders.some(
+    b => b.config && b.config.zeroConfig && b.src.startsWith('api/')
+  );
+  return found ? 'api' : null;
+}
+
+/**
+ * TODO
+ * Replace this function with `config.outputDirectory`
+ */
+function getPublicBuilder(builders: Builder[]): Builder | null {
+  const builder = builders.find(
+    builder =>
+      builder.use === '@now/static' &&
+      /^.*\/\*\*\/\*$/.test(builder.src) &&
+      builder.config &&
+      builder.config.zeroConfig === true
+  );
+
+  return builder || null;
+}
+export function detectOutputDirectory(builders: Builder[]): string | null {
+  // TODO: We eventually want to save the output directory to
+  // builder.config.outputDirectory so it is only detected once
+  const publicBuilder = getPublicBuilder(builders);
+  return publicBuilder ? publicBuilder.src.replace('/**/*', '') : null;
 }
 
 export async function detectBuilders(
@@ -33,11 +81,17 @@ export async function detectBuilders(
   builders: Builder[] | null;
   errors: ErrorResponse[] | null;
   warnings: ErrorResponse[];
+  defaultRoutes: Route[] | null;
+  redirectRoutes: Route[] | null;
 }> {
-  const apiBuilders: Builder[] = [];
   const errors: ErrorResponse[] = [];
   const warnings: ErrorResponse[] = [];
+
+  const apiBuilders: Builder[] = [];
   let frontendBuilder: Builder | null = null;
+
+  const defaultRoutes: Route[] = [];
+  const redirectRoutes: Route[] = [];
 
   const functionError = validateFunctions(options);
 
@@ -46,11 +100,14 @@ export async function detectBuilders(
       builders: null,
       errors: [functionError],
       warnings,
+      defaultRoutes: null,
+      redirectRoutes: null,
     };
   }
 
   const apiMatches = getApiMatches(options);
   const sortedFiles = files.sort(sortFiles);
+  const apiSortedFiles = files.sort(sortFilesBySegmentCount);
 
   // Keep track of functions that are used
   const usedFunctions = new Set<string>();
@@ -73,11 +130,38 @@ export async function detectBuilders(
   let hasNoneApiFiles = false;
   let hasNextApiFiles = false;
 
+  const preDefaultRoutes: Source[] = [];
+  const preDynamicRoutes: Source[] = [];
+
   // API
   for (const fileName of sortedFiles) {
     const apiBuilder = maybeGetApiBuilder(fileName, apiMatches, options);
 
     if (apiBuilder) {
+      const { routeError, defaultRoute, dynamicRoute } = getApiRoute(
+        fileName,
+        apiSortedFiles,
+        options
+      );
+
+      if (routeError) {
+        return {
+          builders: null,
+          errors: [routeError],
+          warnings,
+          defaultRoutes: null,
+          redirectRoutes: null,
+        };
+      }
+
+      if (dynamicRoute) {
+        preDynamicRoutes.push(dynamicRoute);
+      }
+
+      if (defaultRoute) {
+        preDefaultRoutes.push(defaultRoute);
+      }
+
       addToUsedFunctions(apiBuilder);
       apiBuilders.push(apiBuilder);
       continue;
@@ -122,7 +206,13 @@ export async function detectBuilders(
       // We only show this error when there are no api builders
       // since the dependencies of the pkg could be used for those
       errors.push(getMissingBuildScriptError());
-      return { errors, warnings, builders: null };
+      return {
+        errors,
+        warnings,
+        builders: null,
+        redirectRoutes: null,
+        defaultRoutes: null,
+      };
     }
 
     // If `outputDirectory` is an empty string,
@@ -160,6 +250,8 @@ export async function detectBuilders(
       builders: null,
       errors: [unusedFunctionError],
       warnings,
+      redirectRoutes: null,
+      defaultRoutes: null,
     };
   }
 
@@ -181,10 +273,24 @@ export async function detectBuilders(
     }
   }
 
+  const routesResult = mergeRoutes(
+    preDefaultRoutes,
+    preDynamicRoutes,
+    usedOutputDirectory,
+    apiBuilders,
+    frontendBuilder,
+    options
+  );
+
+  defaultRoutes.push(...routesResult.defaultRoutes);
+  redirectRoutes.push(...routesResult.redirectRoutes);
+
   return {
     warnings,
     builders: builders.length ? builders : null,
     errors: errors.length ? errors : null,
+    redirectRoutes,
+    defaultRoutes,
   };
 }
 
@@ -491,42 +597,368 @@ function checkUnusedFunctions(
   return null;
 }
 
-/**
- * TODO
- * Remove this function
- */
-export function getIgnoreApiFilter(optionsOrBuilders: Options | Builder[]) {
-  const possiblePatterns: string[] = getApiMatches().map(b => b.src);
+function getApiRoute(
+  fileName: string,
+  sortedFiles: string[],
+  options: Options
+): {
+  defaultRoute: Source | null;
+  dynamicRoute: Source | null;
+  routeError: ErrorResponse | null;
+} {
+  const conflictingSegment = getConflictingSegment(fileName);
 
-  if (Array.isArray(optionsOrBuilders)) {
-    optionsOrBuilders.forEach(({ src }) => possiblePatterns.push(src));
-  } else if (optionsOrBuilders.functions) {
-    Object.keys(optionsOrBuilders.functions).forEach(p =>
-      possiblePatterns.push(p)
-    );
+  if (conflictingSegment) {
+    return {
+      defaultRoute: null,
+      dynamicRoute: null,
+      routeError: {
+        code: 'conflicting_path_segment',
+        message:
+          `The segment "${conflictingSegment}" occurs more than ` +
+          `one time in your path "${fileName}". Please make sure that ` +
+          `every segment in a path is unique.`,
+      },
+    };
   }
 
-  return (file: string) => {
-    if (!file.startsWith('api/')) {
-      return false;
-    }
+  const occurrences = pathOccurrences(fileName, sortedFiles).filter(
+    name => name !== fileName
+  );
 
-    if (file.includes('/.')) {
-      return false;
-    }
+  if (occurrences.length > 0) {
+    const messagePaths = concatArrayOfText(
+      occurrences.map(name => `"${name}"`)
+    );
 
-    if (file.includes('/_')) {
-      return false;
-    }
+    return {
+      defaultRoute: null,
+      dynamicRoute: null,
+      routeError: {
+        code: 'conflicting_file_path',
+        message:
+          `Two or more files have conflicting paths or names. ` +
+          `Please make sure path segments and filenames, without their extension, are unique. ` +
+          `The path "${fileName}" has conflicts with ${messagePaths}.`,
+      },
+    };
+  }
 
-    if (file.endsWith('.d.ts')) {
-      return false;
-    }
+  const out = createRouteFromPath(
+    fileName,
+    Boolean(options.featHandleMiss),
+    Boolean(options.cleanUrls)
+  );
 
-    if (possiblePatterns.every(p => !(file === p || minimatch(file, p)))) {
-      return false;
-    }
-
-    return true;
+  return {
+    defaultRoute: out.route,
+    dynamicRoute: out.isDynamic ? out.route : null,
+    routeError: null,
   };
+}
+
+// Checks if a placeholder with the same name is used
+// multiple times inside the same path
+function getConflictingSegment(filePath: string): string | null {
+  const segments = new Set<string>();
+
+  for (const segment of filePath.split('/')) {
+    const name = getSegmentName(segment);
+
+    if (name !== null && segments.has(name)) {
+      return name;
+    }
+
+    if (name) {
+      segments.add(name);
+    }
+  }
+
+  return null;
+}
+
+// Takes a filename or foldername, strips the extension
+// gets the part between the "[]" brackets.
+// It will return `null` if there are no brackets
+// and therefore no segment.
+function getSegmentName(segment: string): string | null {
+  const { name } = parsePath(segment);
+
+  if (name.startsWith('[') && name.endsWith(']')) {
+    return name.slice(1, -1);
+  }
+
+  return null;
+}
+
+// Counts how often a path occurs when all placeholders
+// got resolved, so we can check if they have conflicts
+function pathOccurrences(fileName: string, files: string[]): string[] {
+  const getAbsolutePath = (unresolvedPath: string): string => {
+    const { dir, name } = parsePath(unresolvedPath);
+    const parts = joinPath(dir, name).split('/');
+    return parts.map(part => part.replace(/\[.*\]/, '1')).join('/');
+  };
+
+  const currentAbsolutePath = getAbsolutePath(fileName);
+
+  // TODO - We still iterate over each file inside of here
+  return files.reduce((prev: string[], file: string): string[] => {
+    const absolutePath = getAbsolutePath(file);
+
+    if (absolutePath === currentAbsolutePath) {
+      prev.push(file);
+    } else if (partiallyMatches(fileName, file)) {
+      prev.push(file);
+    }
+
+    return prev;
+  }, []);
+}
+
+function joinPath(...segments: string[]) {
+  const joinedPath = segments.join('/');
+  return joinedPath.replace(/\/{2,}/g, '/');
+}
+
+function escapeName(name: string) {
+  const special = '[]^$.|?*+()'.split('');
+
+  for (const char of special) {
+    name = name.replace(new RegExp(`\\${char}`, 'g'), `\\${char}`);
+  }
+
+  return name;
+}
+
+function concatArrayOfText(texts: string[]): string {
+  if (texts.length <= 2) {
+    return texts.join(' and ');
+  }
+
+  const last = texts.pop();
+  return `${texts.join(', ')}, and ${last}`;
+}
+
+// Check if the path partially matches and has the same
+// name for the path segment at the same position
+function partiallyMatches(pathA: string, pathB: string): boolean {
+  const partsA = pathA.split('/');
+  const partsB = pathB.split('/');
+
+  const long = partsA.length > partsB.length ? partsA : partsB;
+  const short = long === partsA ? partsB : partsA;
+
+  let index = 0;
+
+  for (const segmentShort of short) {
+    const segmentLong = long[index];
+
+    const nameLong = getSegmentName(segmentLong);
+    const nameShort = getSegmentName(segmentShort);
+
+    // If there are no segments or the paths differ we
+    // return as they are not matching
+    if (segmentShort !== segmentLong && (!nameLong || !nameShort)) {
+      return false;
+    }
+
+    if (nameLong !== nameShort) {
+      return true;
+    }
+
+    index += 1;
+  }
+
+  return false;
+}
+
+function createRouteFromPath(
+  filePath: string,
+  featHandleMiss: boolean,
+  cleanUrls: boolean
+): { route: Source; isDynamic: boolean } {
+  const parts = filePath.split('/');
+
+  let counter = 1;
+  const query: string[] = [];
+  let isDynamic = false;
+
+  const srcParts = parts.map((segment, i): string => {
+    const name = getSegmentName(segment);
+    const isLast = i === parts.length - 1;
+
+    if (name !== null) {
+      // We can't use `URLSearchParams` because `$` would get escaped
+      query.push(`${name}=$${counter++}`);
+      isDynamic = true;
+      return `([^/]+)`;
+    } else if (isLast) {
+      const { name: fileName, ext } = parsePath(segment);
+      const isIndex = fileName === 'index';
+      const prefix = isIndex ? '\\/' : '';
+
+      const names = [
+        isIndex ? prefix : `${fileName}\\/`,
+        prefix + escapeName(fileName),
+        featHandleMiss && cleanUrls
+          ? ''
+          : prefix + escapeName(fileName) + escapeName(ext),
+      ].filter(Boolean);
+
+      // Either filename with extension, filename without extension
+      // or nothing when the filename is `index`.
+      // When `cleanUrls: true` then do *not* add the filename with extension.
+      return `(${names.join('|')})${isIndex ? '?' : ''}`;
+    }
+
+    return segment;
+  });
+
+  const { name: fileName, ext } = parsePath(filePath);
+  const isIndex = fileName === 'index';
+  const queryString = `${query.length ? '?' : ''}${query.join('&')}`;
+
+  const src = isIndex
+    ? `^/${srcParts.slice(0, -1).join('/')}${srcParts.slice(-1)[0]}$`
+    : `^/${srcParts.join('/')}$`;
+
+  let route: Source;
+
+  if (featHandleMiss) {
+    const extensionless = ext ? filePath.slice(0, -ext.length) : filePath;
+    route = {
+      src,
+      dest: `/${extensionless}${queryString}`,
+      check: true,
+    };
+  } else {
+    route = {
+      src,
+      dest: `/${filePath}${queryString}`,
+    };
+  }
+
+  return { route, isDynamic };
+}
+
+function mergeRoutes(
+  preDefaultRoutes: Source[],
+  preDynamicRoutes: Source[],
+  outputDirectory: string,
+  apiBuilders: Builder[],
+  frontendBuilder: Builder | null,
+  options: Options
+): {
+  defaultRoutes: Route[];
+  redirectRoutes: Route[];
+} {
+  const defaultRoutes: Route[] = [];
+  const redirectRoutes: Route[] = [];
+
+  if (preDefaultRoutes && preDefaultRoutes.length > 0) {
+    if (options.featHandleMiss) {
+      defaultRoutes.push({ handle: 'miss' });
+
+      const extSet = detectApiExtensions(apiBuilders);
+
+      if (extSet.size > 0) {
+        const exts = Array.from(extSet)
+          .map(ext => ext.slice(1))
+          .join('|');
+
+        const extGroup = `(?:\\.(?:${exts}))`;
+
+        if (options.cleanUrls) {
+          redirectRoutes.push({
+            src: `^/(api(?:.+)?)/index${extGroup}?/?$`,
+            headers: { Location: options.trailingSlash ? '/$1/' : '/$1' },
+            status: 308,
+          });
+
+          redirectRoutes.push({
+            src: `^/api/(.+)${extGroup}/?$`,
+            headers: {
+              Location: options.trailingSlash ? '/api/$1/' : '/api/$1',
+            },
+            status: 308,
+          });
+        } else {
+          defaultRoutes.push({
+            src: `^/api/(.+)${extGroup}$`,
+            dest: '/api/$1',
+            check: true,
+          });
+        }
+      }
+
+      if (preDynamicRoutes) {
+        defaultRoutes.push(...preDynamicRoutes);
+      }
+
+      if (preDefaultRoutes.length) {
+        defaultRoutes.push({
+          src: '^/api(/.*)?$',
+          status: 404,
+          continue: true,
+        });
+      }
+    } else {
+      defaultRoutes.push(...preDefaultRoutes);
+
+      if (preDefaultRoutes.length) {
+        defaultRoutes.push({
+          status: 404,
+          src: '^/api(/.*)?$',
+        });
+      }
+    }
+  }
+
+  if (
+    outputDirectory &&
+    frontendBuilder &&
+    !options.featHandleMiss &&
+    frontendBuilder.use === '@now/static'
+  ) {
+    defaultRoutes.push({
+      src: '/(.*)',
+      dest: `/${outputDirectory}/$1`,
+    });
+  }
+
+  return {
+    defaultRoutes,
+    redirectRoutes,
+  };
+}
+
+function sortFilesBySegmentCount(fileA: string, fileB: string): number {
+  const lengthA = fileA.split('/').length;
+  const lengthB = fileB.split('/').length;
+
+  if (lengthA > lengthB) {
+    return -1;
+  }
+
+  if (lengthA < lengthB) {
+    return 1;
+  }
+
+  // Paths that have the same segment length but
+  // less placeholders are preferred
+  const countSegments = (prev: number, segment: string) =>
+    getSegmentName(segment) ? prev + 1 : 0;
+  const segmentLengthA = fileA.split('/').reduce(countSegments, 0);
+  const segmentLengthB = fileB.split('/').reduce(countSegments, 0);
+
+  if (segmentLengthA > segmentLengthB) {
+    return 1;
+  }
+
+  if (segmentLengthA < segmentLengthB) {
+    return -1;
+  }
+
+  return 0;
 }
